+++ docs/lib/menuData.ts
@@ -51,6 +51,7 @@ export default {
         'Banner',
         'Descriptions',
         'Entity',
+        'Empty',
         'Form',
         'Dropdown',
       ],