+++ packages/components/src/ConfigProvider/LocaleProvider/types.ts
@@ -1,8 +1,9 @@
 import { ModalLocale } from '../../Modal/locale';
-// eslint-disable-next-line import/no-cycle
 import { PickerLocale } from '../../DatePicker/generatePicker';
 
 export interface ILocale {
   Modal?: ModalLocale;
   DatePicker?: PickerLocale;
 }
+
+export type Locale = 'en' | 'zh' | 'zh-tw' | 'es';