+++ packages/components/src/ConfigProvider/locales/en.ts
@@ -1,7 +1,5 @@
-// eslint-disable-next-line import/no-cycle
 import { ILocale } from '../LocaleProvider/types';
-// eslint-disable-next-line import/no-cycle
-import PickerLocale from '../../DatePicker/locales/en_US';
+import PickerLocale from '../../DatePicker/locales/en';
 
 const localeValues: ILocale = {
   Modal: {