+++ packages/components/src/DatePicker/locales/es.ts
@@ -1,0 +1,17 @@
+import CalendarLocale from 'rc-picker/lib/locale/es_ES';
+import TimePickerLocale from '../../TimePicker/locales/es';
+import { PickerLocale } from '../generatePicker';
+
+// Merge into a locale object
+const locale: PickerLocale = {
+  lang: {
+    placeholder: 'Seleccionar fecha',
+    rangePlaceholder: ['Fecha inicial', 'Fecha final'],
+    ...CalendarLocale,
+  },
+  timePickerLocale: {
+    ...TimePickerLocale,
+  },
+};
+
+export default locale;